models/inventory-entry/src/inventory-entry-draft/presets/sample-data-b2c-lifestyle/sku-ctss-098.ts
import {
  ProductVariantDraft,
  type TProductVariantDraft,
} from '@commercetools-test-data/product';
import type { TInventoryEntryDraftBuilder } from '../../../types';
import * as InventoryEntryDraft from '../../index';

const variant = ProductVariantDraft.presets.sampleData
  .canelaThreeSeaterSofa01()
  .build<TProductVariantDraft>();

const skuCtss098 = (): TInventoryEntryDraftBuilder =>
  InventoryEntryDraft.presets.empty().sku(variant.sku!).quantityOnStock(1000);
export default skuCtss098;
